avoiding uint32 to int conversion

+++ src/fheroes2/battle/battle_interface.cpp
@@ -6100,8 +6100,8 @@ void Battle::PopupDamageInfo::SetSpellAttackInfo( const Cell * cell, const Unit 
         return;
     }
 
-    const uint32_t spellPoints = hero ? hero->GetPower() : DEFAULT_SPELL_DURATION;
-    const uint32_t spellDamage = defender->CalculateDamage( spell, spellPoints, hero, 0 /* targetInfo damage */, true /* ignore defending hero */ );
+    const int spellPoints = hero ? hero->GetPower() : DEFAULT_SPELL_DURATION;
+    const uint32_t spellDamage = defender->CalculateDamage( spell, (uint32_t) spellPoints, hero, 0 /* targetInfo damage */, true /* ignore defending hero */ );
 
     _redraw = true;
     _minDamage = spellDamage;
